refactor pw class
