+++ app/src/main/java/com/sdp/swiftwallet/presentation/signIn/ForgotPasswordActivity.java
@@ -1,14 +1,18 @@
 package com.sdp.swiftwallet.presentation.signIn;
 
+import android.app.Activity;
+import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
+import com.google.firebase.auth.FirebaseAuth;
 import com.sdp.cryptowalletapp.R;
-import com.sdp.swiftwallet.data.repository.FirebaseAuthImpl;
+import com.sdp.swiftwallet.common.FirebaseUtil;
 import com.sdp.swiftwallet.domain.repository.ClientAuth;
 import org.jetbrains.annotations.NotNull;
 
@@ -17,9 +21,11 @@ public class ForgotPasswordActivity extends AppCompatActivity {
   private ClientAuth dbClient;
   private String country;
   private String countryCodeLanguage;
+  private FirebaseAuth mAuth;
+
 
   private EditText emailView;
-
+  private static final String RESET_PASSWORD_TAG = "RESET_PASSWORD_TAG";
 
   @Override
   public void onCreate(Bundle savedInstanceState) {
@@ -27,6 +33,9 @@ public class ForgotPasswordActivity extends AppCompatActivity {
     super.onCreate(savedInstanceState);
     //Set the content
     setContentView(R.layout.activity_reset_password);
+
+    //Sets up the db
+    mAuth = FirebaseUtil.getAuth();
 
     // Get the contents of the email fieldemailView = findViewById(R.id.emailField);
     country = "en";
@@ -42,9 +51,8 @@ public class ForgotPasswordActivity extends AppCompatActivity {
       }
     });
 
-    dbClient = new FirebaseAuthImpl();
     //Hardcoded, to be changed
-    dbClient.setLanguage("en", "en_gb");
+    setLanguage("en", "en_gb");
   }
 
 
@@ -56,7 +64,7 @@ public class ForgotPasswordActivity extends AppCompatActivity {
   public void checkAndSend(@NotNull String email){
 
     if (isEmailValid(email)) {
-      dbClient.sendPasswordResetEmail(email, country, countryCodeLanguage, this);
+      sendPasswordResetEmail(email, country, countryCodeLanguage, this);
     } else {
       Toast.makeText(this,
           "Erro, please check again the format of the email", Toast.LENGTH_LONG).show();
@@ -64,6 +72,27 @@ public class ForgotPasswordActivity extends AppCompatActivity {
 
   }
 
+  /**
+   * Sends a password reset email to the user
+   * @param email email that must previously have been checked
+   * @param country country code, checked by the check language function
+   * @param countryCode country code for language, checked by the check language function
+   */
+  public void sendPasswordResetEmail(String email, String country, String countryCode, Activity from){
+    setLanguage(country, countryCode);
+    Log.d(RESET_PASSWORD_TAG, "Trying to send a confirmation email");
+    mAuth.sendPasswordResetEmail(email)
+        .addOnSuccessListener( a -> {
+          Log.d(RESET_PASSWORD_TAG, "Password successfully sent on \n"+email);
+          Toast.makeText(from, "Password successfully sent on \n"+email, Toast.LENGTH_SHORT).show();
+          //Start again login activity if successful
+          Intent nextIntent = new Intent(from, LoginActivity.class);
+          from.startActivity(nextIntent);
+        }).addOnFailureListener( a -> {
+      Log.d(RESET_PASSWORD_TAG, "Something went wrong, please enter a valid email \n"+email);
+      Toast.makeText(from, "Reset error, please correct your email! \n"+email, Toast.LENGTH_SHORT).show();
+    });
+  }
 
   /**
    * Check if email is correctly formatted, might become a function called everywhere
@@ -78,6 +107,14 @@ public class ForgotPasswordActivity extends AppCompatActivity {
     return true;
   }
 
-
+  /**
+   * Sets up language for authentication
+   * @param country country (format "fr=France", ...)
+   * @param countryLanguage country language (format "en_gb"= UK british)
+   */
+  public void setLanguage(String country, String countryLanguage){
+    mAuth.setLanguageCode(country);
+    mAuth.setLanguageCode(countryLanguage);
+  }
 
 }
